Move dashboard state into module
