+++ src/modules/dashboard/state/AllRoles.ts
@@ -2,13 +2,7 @@ import { Map as ImmutableMap } from 'immutable';
 import { COLONY_ROLES } from '@colony/colony-js-client';
 
 import { Address } from '~types/index';
-import { FetchableDataRecord } from '../../FetchableData';
-
-export type UserRolesObject = { [role: string]: boolean };
-
-export type DomainRolesObject = { [userAddress: string]: UserRolesObject };
-
-export type ColonyRolesObject = { [domainId: number]: DomainRolesObject };
+import { FetchableDataRecord } from '~immutable/index';
 
 export type ColonyRolesMap = ImmutableMap<
   number,
