+++ src/modules/dashboard/state/index.ts
@@ -1,23 +1,23 @@
 import { List as ListType, Map as ImmutableMap, Record } from 'immutable';
 
 import { Address } from '~types/index';
+import {
+  FetchableDataRecord,
+  TaskDraftId,
+  TaskRecord,
+  TaskFeedItemRecord,
+  TaskMetadataRecord,
+} from '~immutable/index';
+
 import { AllColoniesRecord } from './AllColonies';
 import { AllDomainsMap } from './AllDomains';
 import { AllRolesMap } from './AllRoles';
 import { AllTokensMap } from './AllTokens';
-import { FetchableDataRecord } from '../../FetchableData';
-import { TaskCommentRecord } from '../../TaskComment';
-import { TaskDraftId, TaskRecord } from '../../Task';
-import { TaskFeedItemRecord } from '../../TaskFeedItem';
-import { TaskMetadataRecord } from '../../TaskMetadata';
 
 export * from './AllColonies';
 export * from './AllTokens';
 export * from './AllDomains';
 export * from './AllRoles';
-
-export type CommentsList = ListType<TaskCommentRecord>;
-export type AllCommentsMap = ImmutableMap<TaskDraftId, CommentsList>;
 
 export type TaskMetadataMap = ImmutableMap<TaskDraftId, TaskMetadataRecord>;
 
@@ -38,10 +38,9 @@ export type TaskFeedItemsMap = ImmutableMap<
 
 export type DashboardStateProps = {
   allColonies: AllColoniesRecord;
-  allComments: AllCommentsMap;
   allDomains: AllDomainsMap;
   allRoles: AllRolesMap;
-  allTaskMetadata: AllTaskMetadataMap;
+  taskMetadata: AllTaskMetadataMap;
   allTokens: AllTokensMap;
   taskFeedItems: TaskFeedItemsMap;
   tasks: TasksMap;
@@ -49,10 +48,9 @@ export type DashboardStateProps = {
 
 export class DashboardStateRecord extends Record<DashboardStateProps>({
   allColonies: new AllColoniesRecord(),
-  allComments: ImmutableMap(),
   allDomains: ImmutableMap(),
   allRoles: ImmutableMap(),
-  allTaskMetadata: ImmutableMap(),
+  taskMetadata: ImmutableMap(),
   allTokens: ImmutableMap(),
   taskFeedItems: ImmutableMap(),
   tasks: ImmutableMap(),
