+++ src/types/index.ts
@@ -1,11 +1,10 @@
-// eslint-disable-next-line import/no-cycle
 export * from '../lib/ColonyManager/types';
-// eslint-disable-next-line import/no-cycle
 export * from '../data/types';
 export * from '../lib/database/types';
 export * from './keyTypes';
 export * from './TransactionReceipt';
 export * from './strings';
+export * from './roles';
 
 export type WithKey = {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
