+++ src/types/roles.ts
@@ -1,0 +1,5 @@
+export type UserRolesObject = { [role: string]: boolean };
+
+export type DomainRolesObject = { [userAddress: string]: UserRolesObject };
+
+export type ColonyRolesObject = { [domainId: number]: DomainRolesObject };
